Refactoring
* Added a findByHash() method in the tag, place, link and media
DAOs. This method takes a list of hash values as a parameter
and uses them to return the entities with those values in their "hash"
column

* Moved the URL slug and MD5 hash generation from EntityDTO
to SwiftRiverUtils

* Removed @Transactional annotations from the DAO classes. These
should be in the service classes

* Added two new entities; RiverCollaborator and BucketCollaborator
to represent the river_ and bucket_ collaborators tables respectively. These
tables have additional columns thereby necessitating this