Delete river drop from bucket
* Separate endpoints for deleting BucketDrop and RiverDrops
from a bucket. Requests to delete a RiverDrop from a bucket
bear the signature: "/buckets/:id/river/drops/:drop_id". The alternative
is to include the source of the drop i.e. river or bucket in the request body
but the RFC2616 recommends that servers ignore the request body
for DELETE since there are no "defined semantics" for DELETE entity bodies