RESTful URI for deleting bucket drops
* Removed additional endpoint for removing river drops from a bucket
The URL signature was not in conformance with REST patterns

* Replaced the Drop#buckets  property with Drop#bucketDrops; a collection
of BucketDrop associated with the drop entity. This is a transient property and is
set by JpaDropDao#populateBuckets. On the GetDropDTO end, the bucketDrops
property has been mapped to the "buckets" JSON property.  The rest of the property
mapping is in dozer-bean-mappings.xml

* When a river/bucket drop is deleted, delete the associated read drops